state_machine: Add state machine and states
Add initial implementation of a state machine and states
to run update:

1. State machine to initialize update states and execute them.
2. States for each update stage.

It allows splitting a complex implementation with lots of "IFs" into
isolated/encapsulated functional blocks that performs a single function.

In addition to that it allows to run the complex "update" CLI command
and the granular CLI commands (fetch/install/start) in uniform and
consistent way.

Signed-off-by: Mike Sul <mike.sul@foundries.io>