e2e-tests: Add pytest e2e test script
This script is a copy from the one used in aktualizr-lite, where some
the flow was adapted when fioup is being tested.

The tests interact with an actual factory, with pre-created targets.
Not only the state of the virtual device is verified, but also the
generated events send back to the device gateway.

Not all tests originally intented for aktualizr-lite are supported.
The following tests are working correctly with fioup (and can be
specified with the -k flag in pytest):
- 'test_incremental_updates[False-True-False'
- 'test_incremental_updates[False-False-True'
- 'test_random_updates[False-True-False'
- 'test_random_updates[False-False-True'
- 'test_apps_selection'

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>