register: Fix logic for installing docker cred helper
This logic was copied too well from what we do in Fioctl. In Fioctl we
have a different problem; a credential helper needs to be installed and
the user probably needs to run `fioctl configure-docker` with `sudo` to
allow for that. However, it then needs to *configure* the cred helper
for the actual user, `SUDO_USER`.

In the case of fioup, we are wanting to configure as the root user.

In the advanced scenario of running as non-root, this code will contiue
to work correctly.

Signed-off-by: Andy Doan <andy@foundries.io>