check: Use full current target when doing sync update
If current target is still listed in targets.json, use it as ToTarget
when doing a sync update. This ensures that all apps that are part of
the current target are considered when determining which apps need to be
installed.
Otherwhise, only apps that were part of the last successful update would
be considered.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>