start: Retry completion of successfully started update
Update completion (uninstalling and removing dangling blobs) may fail
under unexpected condition after successful update start.
It makes sense to retry the completion in such the case, and if it fails
again then retry without dangling blobs pruning.
If it still fails then do the force update completion, so the update is
considered as a successful and a user can do new updates.

Signed-off-by: Mike Sul <mike.sul@foundries.io>