config: Add storage usage watermark parameter
Add a storage usage watermark parameter to the configuration.
The parameter indicates the maximum percentage out of the overall
storage capacity that can be used by apps managed by `fioup`.

The parameter key is `pacman.storage_watermark` is the same as in aklite
to have compatible configuration for both use-cases.

Signed-off-by: Mike Sul <mike.sul@foundries.io>